simplified version of mathod failure handler
